requirements.txt
Babel==2.7.0
Django==2.2.7
django-phonenumber-field==3.0.1
phonenumbers==8.11.0
pytz==2019.3
coverage==4.5.4
